UPD: refactor save, update, delete and approve methods according reviewer requests

+++ src/main/java/ru/practicum/shareit/booking/service/BookingServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.practicum.shareit.booking.service;
 
 
+import jakarta.validation.ValidationException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,35 +14,28 @@ import ru.practicum.shareit.booking.enums.State;
 import ru.practicum.shareit.booking.enums.Status;
 import ru.practicum.shareit.booking.mapper.BookingMapper;
 import ru.practicum.shareit.booking.repository.BookingRepository;
+import ru.practicum.shareit.common.CommonChecker;
 import ru.practicum.shareit.exception.*;
 import ru.practicum.shareit.item.entity.Item;
-import ru.practicum.shareit.item.repository.ItemRepository;
 import ru.practicum.shareit.user.entity.User;
-import ru.practicum.shareit.user.repository.UserRepository;
 
+import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.stream.Collectors;
 
 @Slf4j
 @Service
-public class BookingServiceImpl implements BookingService {
+public class BookingServiceImpl extends CommonChecker implements BookingService {
     private BookingRepository bookingRepository;
-    private ItemRepository itemRepository;
-    private UserRepository userRepository;
     private BookingMapper mapper;
 
     @Autowired
     public BookingServiceImpl(BookingRepository bookingRepository,
-                              ItemRepository itemRepository,
-                              UserRepository userRepository,
                               BookingMapper mapper) {
         this.bookingRepository = bookingRepository;
-        this.itemRepository = itemRepository;
-        this.userRepository = userRepository;
         this.mapper = mapper;
     }
-
 
     @Override
     @Transactional
@@ -50,12 +44,11 @@ public class BookingServiceImpl implements BookingService {
         Item item = checkItemAndReturn(bookingDto.getItemId());
         User user = checkUserAndReturn(userId);
 
-        if (!item.getAvailable()) {
-            throw new ItemAvailabilityException("Item is not available");
-        }
-        if (item.getOwner().getId().equals(userId)) {
-            throw new ItemAvailabilityException("Can not book own item");
-        }
+        LocalDateTime start = bookingDto.getStart();
+        LocalDateTime end = bookingDto.getEnd();
+
+        validateBookingDates(start, end);
+        validateItemAvailability(userId, item);
 
         Booking booking = mapper.toBooking(bookingDto, item, user);
         bookingRepository.save(booking);
@@ -66,25 +59,60 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     @Transactional
-    public BookingDto updateBooking(BookingUpdateDto bookingDto) {
-        log.debug("Update booking request received. Booking id: {}", bookingDto.getId());
+    public BookingDto updateBooking(Long userId, BookingUpdateDto bookingUpdateDto) {
+        log.debug("Update booking request received. Booking id: {}", bookingUpdateDto.getId());
 
-        Booking updBooking = mapper.updateBookingFromDto(bookingDto, checkBookingAndReturn(bookingDto.getId()));
+        Booking updBooking = checkBookingAndReturn(bookingUpdateDto.getId());
+        boolean isBooker = updBooking.getBooker().getId().equals(userId);
+
+        if (!isBooker) {
+            log.warn("Access denied: invalid user");
+            throw new BookingAccessException("Booker can update booking only");
+        }
+
+        // статус
+        if (bookingUpdateDto.getStatus() != null) {
+            if ((updBooking.getStatus() == Status.WAITING || updBooking.getStatus() == Status.APPROVED) &&
+                    bookingUpdateDto.getStatus() == Status.CANCELLED) {
+                updBooking.setStatus(Status.CANCELLED);
+            } else {
+                log.warn("Invalid status update attempt. Current status: {}, new status: {}",
+                        updBooking.getStatus(), bookingUpdateDto.getStatus());
+                throw new BookingAccessException("Booking can be cancelled in statuses 'WAITING' or 'APPROVED' only");
+            }
+        }
+
+        // время
+        if (bookingUpdateDto.getStart() != null || bookingUpdateDto.getEnd() != null) {
+            if (!updBooking.getStatus().equals(Status.WAITING)) {
+                log.warn("Invalid time to update booking. Booking id: {}",
+                        updBooking.getId());
+                throw new BookingAccessException("Booking dates can only be updated in 'WAITING' status only");
+            }
+            validateBookingDates(bookingUpdateDto.getStart(), bookingUpdateDto.getEnd());
+            mapper.updateBookingFromDto(bookingUpdateDto, updBooking);
+        }
+
         bookingRepository.save(updBooking);
-
-        log.debug("Updating successful!");
+        log.debug("Updating successful! Booking id: {}", updBooking.getId());
         return mapper.toDto(updBooking);
     }
 
     @Override
     @Transactional
-    public BookingDto deleteBooking(Long bookingId) {
+    public BookingDto deleteBooking(Long userId, Long bookingId) {
         log.debug("Delete booking request received. Booking id: {}", bookingId);
 
         Booking booking = checkBookingAndReturn(bookingId);
-        bookingRepository.delete(booking);
+        boolean isBooker = booking.getBooker().getId().equals(userId);
 
-        log.debug("Deleting successful!");
+        if (isBooker) {
+            bookingRepository.delete(booking);
+            log.debug("Deleting successful!");
+        } else {
+            log.debug("Check permissions. User is not a booker");
+            throw new BookingAccessException("Only the booker can cancel the booking");
+        }
         return mapper.toDto(booking);
     }
 
@@ -98,7 +126,7 @@ public class BookingServiceImpl implements BookingService {
         User owner = checkUserAndReturn(booking.getItem().getOwner().getId());
 
         if (!booking.getBooker().getId().equals(userId) && !owner.getId().equals(userId)) {
-            throw new ItemOwnerException("Not enough permissions: booker and owner can see booking only");
+            throw new BookingAccessException("Not enough permissions: booker and owner can see booking only");
         }
 
         return mapper.toDto(booking);
@@ -133,6 +161,7 @@ public class BookingServiceImpl implements BookingService {
     @Transactional(readOnly = true)
     public Collection<BookingDto> findAllBookingsByOwnerAndState(Long userId, String state) {
         log.debug("Get all bookings of owner request received. Owner id: {}. State: {}", userId, state);
+
         State currState = State.valueOf(state);
         User user = checkUserAndReturn(userId);
         Collection<Booking> bookings;
@@ -157,6 +186,7 @@ public class BookingServiceImpl implements BookingService {
     @Transactional
     public BookingDto approveBooking(Long bookingId, Long userId, Boolean approved) {
         log.debug("Approve booking request received. Booking id: {}", bookingId);
+
         Booking booking = checkBookingAndReturn(bookingId);
         Item item = checkItemAndReturn(booking.getItem().getId());
 
@@ -164,32 +194,30 @@ public class BookingServiceImpl implements BookingService {
             throw new CommentIncorrectTimeException("Booking can be approved by owner only");
         }
         if (!booking.getStatus().equals(Status.WAITING)) {
-            throw new ItemAvailabilityException("Booking can be approved in \"WAITING\" status only");
+            throw new ItemAvailabilityException("Booking can be approved in 'WAITING' status only");
         }
 
         booking.setStatus(approved ? Status.APPROVED : Status.REJECTED);
+        bookingRepository.save(booking);
 
         return mapper.toDto(booking);
     }
 
-    private Booking checkBookingAndReturn(Long bookingId) {
-        log.debug("Checking booking");
-
-        return bookingRepository.findById(bookingId).orElseThrow(() ->
-                new NoSuchBookingException("Incorrect booking id: " + bookingId));
+    private void validateBookingDates(LocalDateTime start, LocalDateTime end) {
+        if (!start.isBefore(end) || start.equals(end)) {
+            log.warn("Check booking time: start - {}, end - {}", start, end);
+            throw new ValidationException("Incorrect booking time: the start must be early than the end");
+        }
     }
 
-    private Item checkItemAndReturn(Long itemId) {
-        log.debug("Checking item");
-
-        return itemRepository.findById(itemId).orElseThrow(() ->
-                new NoSuchItemException("Incorrect user id: " + itemId));
-    }
-
-    private User checkUserAndReturn(Long userId) {
-        log.debug("Checking user");
-
-        return userRepository.findById(userId).orElseThrow(() ->
-                new NoSuchUserException("Incorrect user id: " + userId));
+    private void validateItemAvailability(Long userId, Item item) {
+        if (!item.getAvailable()) {
+            log.warn("Check availability of item. Item id: {}", item.getId());
+            throw new ItemAvailabilityException("Item is not available");
+        }
+        if (item.getOwner().getId().equals(userId)) {
+            log.warn("Check booker. Booker id: {}", userId);
+            throw new ItemAvailabilityException("Cannot book own item");
+        }
     }
 }
